requirements.txt
requests>=2.28.0
google-generativeai>=0.3.0